// frontend/src/api/generated/models/DeliveryRecord.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeliveryRecord = {
    id: number;
    userId: number;
    bookMd5: string;
    deviceType: string;
    status: string;
    sentAt?: string | null;
    error?: string | null;
    createdAt: string;
};